requirements.txt
customtkinter
SpeechRecognition
pyttsx3
python-docx
openpyxl
pyaudio
